Use a table to define the redis module.
With this change is now possible to require redis-lua by assigning the
returned module to a local variable like the following line:

    local redis = require 'redis'

Starting with this release this is the preferred way to require redis-lua.
The `Redis` global name is still available for backwards compatibility but
it will be removed in the next major version.