Do not hang when first argument of a command is a nil value.
When passing nil values in command arguments they are now converted to
empty strings explicitly, on the other hand when nil is the last value
then it will not be passed to the serializer due to how Lua works with
nil values in tables, but at least the client will not hang and Redis
will return a -ERR complaining about the wrong number of arguments for
the given command.

Fixes #29.